--- src/main/kotlin/gomushin/backend/core/jwt/infrastructrue/Type.kt
@@ -1,6 +1,0 @@
-package gomushin.backend.core.jwt.infrastructrue
-
-enum class Type {
-    ACCESS, REFRESH
-}
-
