+++ src/main/kotlin/gomushin/backend/core/jwt/infrastructure/Type.kt
@@ -1,0 +1,6 @@
+package gomushin.backend.core.jwt.infrastructure
+
+enum class Type {
+    ACCESS, REFRESH
+}
+
